requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4
blinker==1.7.0
blis==0.7.11
branca==0.7.1
cachetools==5.3.2
catalogue==2.0.10
catboost==1.2.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cog==0.9.4
colorama==0.4.6
comtypes==1.3.1
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.1
cssselect2==0.7.0
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
decouple==0.0.7

dm-tree==0.1.8
executing==2.0.1
Faker==19.13.0
fastapi==0.98.0
filelock==3.13.1
fiona==1.9.6
Flask==3.0.0
Flask-Mail==0.9.1
flatbuffers==23.5.26
fonttools==4.47.2
frozendict==2.4.0
fsspec==2024.2.0
gast==0.5.4
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googlesearch-python
graphviz==0.20.1
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httptools==0.6.1
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
module-name==0.6.0
monkeylearn==3.6.0
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.10
namex==0.0.7
networkx==3.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
oscrypto==1.3.0
outcome==1.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
passlib==1.7.4
peewee==3.17.0
pillow==10.2.0
plotly==5.18.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.1
pypdf==4.0.1
pypiwin32==223
pypng==0.20220715.0
pyproj==3.6.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-speech-features==0.6
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
rdt==1.10.1
referencing==0.33.0
regex==2023.12.25
reportlab==4.0.9

rich==13.7.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.1
selenium==4.14.0
shapely==2.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
structlog==24.1.0
svglib==1.5.1
sympy==1.12
termcolor==2.4.0
thinc==8.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.1.30
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
transformers==4.37.2
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.20
uritools==4.0.2
urllib3==2.2.0
uvicorn==0.27.0.post1
validators==0.22.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
whois==0.9.13
wrapt==1.14.1
wsproto==1.2.0
xhtml2pdf==0.2.14
xyzservices==2023.10.1
zipp==3.17.0
